refactor: modernize codebase with C++20/23 features and clean up codebase
- Flatten Chunk block storage to a 1D std::array and utilize C++23 std::mdspan for 3D indexing.
- Use C++17 Class Template Argument Deduction (CTAD) for std::array instantiations.
- Adopt structured bindings for map iterations and queue family retrieval.
- Simplify high-resolution clock delta calculations and clean up redundant static_casts.
- Remove unused VkDevice parameter from MegaBuffer constructor.
- Add CMakePresets.json for developer build presets and vendor/.editorconfig for formatting.

+++ src/input/GLFWInput.cpp
@@ -14,13 +14,13 @@ namespace voxl
 
     bool GLFWInput::isKeyPressedImpl(KeyCode keyCode) const
     {
-        const auto state = glfwGetKey(m_window, static_cast<int>(std::to_underlying(keyCode)));
+        const auto state = glfwGetKey(m_window, std::to_underlying(keyCode));
         return state == GLFW_PRESS || state == GLFW_REPEAT;
     }
 
     bool GLFWInput::isMouseButtonPressedImpl(MouseCode mouseCode) const
     {
-        const auto state = glfwGetMouseButton(m_window, static_cast<int>(std::to_underlying(mouseCode)));
+        const auto state = glfwGetMouseButton(m_window, std::to_underlying(mouseCode));
         return state == GLFW_PRESS;
     }
